% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cooccurence_analysis.R
\name{Binary.LR.Diffexp}
\alias{Binary.LR.Diffexp}
\title{Find differentially expressed genes between Spatial spots Double positive and Double negative for an LR pair}
\usage{
Binary.LR.Diffexp(SeuratObj, Enriched.LRs.df, COI, Condition)
}
\arguments{
\item{SeuratObj}{Spatial Seurat Object}

\item{Enriched.LRs.df}{data.frame of selected LR, output of Diff.cooc.LRs function}

\item{COI}{Composition Cluster of Interest}

\item{Condition}{vector specifying conditions of interest from orig.ident metadata column}
}
\value{
list of Differentially expressed genes between Spatial spots Double positive and Double negative for LRs in Enriched.LRs.df
}
\description{
This function Finds differentially expressed genes between Spatial spots Double positive and Double negative for an LR pair
}
