% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composition_clustering.R
\name{Composition.cluster.k}
\alias{Composition.cluster.k}
\title{Define the number of composition clusters}
\usage{
Composition.cluster.k(Celltype_deconvolved_probs, max_k)
}
\arguments{
\item{Celltype_deconvolved_probs}{probability matrix of deconvoluted spatial spots. Rows corresponds to spatial spots, columns corresponds to cell types. can be calculated with or without single-cell reference}

\item{max_k}{Maximum number of expected clusters to iterate}
}
\value{
an elbow plot of the k values to decide for the best k
}
\description{
This function iterates through defined number of clusters and plot the total withinss of kmeans clustering
}
