% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composition_clustering.R
\name{Composition.cluster}
\alias{Composition.cluster}
\title{Clustering spatial spots based on their deconvoluted celltype composition}
\usage{
Composition.cluster(spatial.object, Celltype_deconvolved_probs, k)
}
\arguments{
\item{spatial.object}{processed seurat object of the spatial sample with all annotation}

\item{Celltype_deconvolved_probs}{probability matrix of deconvoluted spatial spots. Rows corresponds to spatial spots, columns corresponds to cell types. can be calculated with or without single-cell reference}

\item{k}{number of expected clusters}
}
\value{
a seurat spatial object with an added column to metadata called CompositionCluster_CC
}
\description{
This function clusters spatial spots based on their deconvoluted celltype composition
}
