% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composition_clustering.R
\name{Composition_cluster_enrichedCelltypes}
\alias{Composition_cluster_enrichedCelltypes}
\title{Defining driver celltypes of each composition cluster}
\usage{
Composition_cluster_enrichedCelltypes(
  Spatial.object,
  COI,
  Celltype_deconvolved_probs
)
}
\arguments{
\item{COI}{Composition Cluster of Interest}

\item{Celltype_deconvolved_probs}{probability matrix of deconvoluted spatial spots. Rows corresponds to spatial spots, columns corresponds to cell types. can be calculated with or without single-cell reference}

\item{spatial.object}{processed seurat object of the spatial sample with all annotation}
}
\value{
a boxplot of cell type probabilities for cluster of interest
}
\description{
This function Defining driver celltypes of each composition cluster
}
