% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composition_clustering.R
\name{Composition_cluster_group_pct}
\alias{Composition_cluster_group_pct}
\title{Defining Condition-specific and Condition-common composition clusters}
\usage{
Composition_cluster_group_pct(Spatial.object, Cluster_Column, Condition)
}
\arguments{
\item{Cluster_Column}{Column from Metadata that contains Composition cluster labels}

\item{Condition}{Column from Metadata that contains Condition labels}

\item{spatial.object}{processed seurat object of the spatial sample with all annotation}
}
\value{
a list containing Condition-specific composition clusters, Condition-common composition clusters and ggplot of the percentages
}
\description{
This function provides Defining Condition-specific and Condition-common composition clusters
}
