% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composition_clustering.R
\name{Composition_cluster_umap}
\alias{Composition_cluster_umap}
\title{Umap of spatial spots based on their cell type composition profile}
\usage{
Composition_cluster_umap(spatial.object, Celltype_deconvolved_probs)
}
\arguments{
\item{spatial.object}{processed seurat object of the spatial sample with all annotation}

\item{Celltype_deconvolved_probs}{probability matrix of deconvoluted spatial spots. Rows corresponds to spatial spots, columns corresponds to cell types. can be calculated with or without single-cell reference}
}
\value{
a list containing UMAP coordinates of spots, ggplot of umap color coded by comosition clusters, ggplot of umap with deconvolution pie charts
}
\description{
This function provides umap coordinates and plots of spatial spots based on celltype composition profiles
}
