% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cooccurence_analysis.R
\name{Diff.cooc.LRs}
\alias{Diff.cooc.LRs}
\title{Find LRs that are significantly cooccuring in one group and not in the other group}
\usage{
Diff.cooc.LRs(Enriched.LRs.G1, Enriched.LRs.G2, G1.max.pth, G2.min.pth)
}
\arguments{
\item{Enriched.LRs.G1}{Output of Enriched.LRs function for Group1}

\item{Enriched.LRs.G2}{Output of Enriched.LRs function for Group2}

\item{G1.max.pth}{Max pvalue threshold for significancy levels of cooccuring LRs in Group1}

\item{G2.min.pth}{Min pvalue threshold for non-significancy levels of cooccuring LRs in Group2}
}
\value{
List of LRs enriched in Group1 and not in Group2
}
\description{
This function Find LRs that are significantly cooccuring in one group and not in the other group
}
