% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Differentially_activated_networks.R
\name{DiffNet.Celltype}
\alias{DiffNet.Celltype}
\title{Differentially activated cell-type network between two conditions}
\usage{
DiffNet.Celltype(spatial.object, Celltype_deconvolved_probs, COI1, COI2)
}
\arguments{
\item{spatial.object}{processed seurat object of the spatial sample with all annotation}

\item{Celltype_deconvolved_probs}{probability matrix of deconvoluted spatial spots. Rows corresponds to spatial spots, columns corresponds to cell types. can be calculated with or without single-cell reference}

\item{COI1}{First Composition cluster of interest for the differential celltype interaction network analysis}

\item{COI2}{Second Composition cluster of interest for the differential celltype interaction network analysis}
}
\value{
a list containing the estimated difference of condition specific precision matrices. celltype pairs diff activity scores, The estimated differential network over only the connected nodes
}
\description{
This function calculates Differentially activated cell type network between two conditions
}
