% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Differentially_activated_networks.R
\name{DiffNet.LR}
\alias{DiffNet.LR}
\title{Differentially activated ligand-receptor network between two conditions}
\usage{
DiffNet.LR(spatial.object, LR.list, COI1, Cond1, COI2, Cond2)
}
\arguments{
\item{spatial.object}{processed seurat object of the spatial sample with all annotation}

\item{LR.list}{list of interested ligands and receptors to be explored. Common genes with spatial expression matrix}

\item{COI1}{First Composition cluster of interest for the differential activated l-R network analysis}

\item{Cond1}{The consition that COI1 spots would be subsetted from.}

\item{COI2}{Second Composition cluster of interest for the differential activated l-R network analysis}

\item{Cond2}{The consition that COI2 spots would be subsetted from.}
}
\value{
a list containing the estimated difference of condition specific precision matrices. L-R pairs diff activity scores, The estimated differential network over only the connected nodes
}
\description{
This function calculates Differentially activated ligand-receptor network between two conditions
}
