% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cooccurence_analysis.R
\name{Enriched.LRs}
\alias{Enriched.LRs}
\title{Calculate sigificantly cooccuring Ligand-receptor pairs in composition clusters of interest}
\usage{
Enriched.LRs(
  spatial.object,
  COI,
  Condition,
  LR.list,
  LR.pairs,
  exp.th,
  corr.th
)
}
\arguments{
\item{spatial.object}{processed seurat object of the spatial sample with all annotation}

\item{COI}{vector including Composition clusters of interest of spatial spots which we would like to calculate significant Ligand-receptor cooccurences in}

\item{Condition}{vector including conditions from orig.ident to select COI from}

\item{LR.list}{list of interested ligands and receptors to be explored. Common genes with spatial expression matrix}

\item{LR.pairs}{list of interested ligands and receptors pairs, separated by Underline}

\item{exp.th}{expression threshold for raw counts of ligand-receptors to convert gene expression matrix to binary presence/absence matrix}

\item{corr.th}{correlation threshold between ligand and receptor pair expression}
}
\value{
a list of significantly cooccuring and correlating ligand annd receptors in composition cluster of interest and result table of cooccurence analysis
}
\description{
This function calculates the significant cooccuring Ligand-Receprot pairs based on Ligand-receptor expression matrix
}
