% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cooccurence_analysis.R
\name{plot.celltype.cooccurence}
\alias{plot.celltype.cooccurence}
\title{Function for producing a heatmap co-occurrence visualization.}
\usage{
\method{plot}{celltype.cooccurence}(Cooccure.resulst)
}
\arguments{
\item{Cooccure.resulst}{Output of spatial.celltype.cooccurence funcion}
}
\value{
heatmap plot of cooccurence results
}
\description{
Heatmap visualization of the pairwise species associations revealed by a cooccur analysis
}
