% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cooccurence_analysis.R
\name{SankeyPlot.Diff.LRs}
\alias{SankeyPlot.Diff.LRs}
\title{Plot group specific LRs}
\usage{
SankeyPlot.Diff.LRs(Group1.enrcihed.LRs, Group2.enrcihed.LRs)
}
\arguments{
\item{Group1.enrcihed.LRs}{data.frame of selected LRs with "pair" column}

\item{Group2.enrcihed.LRs}{data.frame of selected LRs with "pair" column}
}
\value{
Chord Diagram of LR interactions
}
\description{
This function Plots group specific LRs
}
