% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cooccurence_analysis.R
\name{spatial.celltype.cooccurence}
\alias{spatial.celltype.cooccurence}
\title{Calculate sigificantly cooccuring celltypes in composition clusters of interest}
\usage{
spatial.celltype.cooccurence(
  spatial.object,
  deconv.prob.mat,
  COI,
  Condition,
  prob.th
)
}
\arguments{
\item{spatial.object}{processed seurat object of the spatial sample with all annotation}

\item{deconv.prob.mat}{probability matrix of deconvoluted spatial spots. Rows corresponds to spatial spots, columns corresponds to cell types. can be calculated with or without single-cell reference}

\item{COI}{Composition cluster of interest, a vector of Cluster IDs, of spatial spots which we would like to calculate significant cell type cooccurences in}

\item{Condition}{Condition of interest,a vector of Conditions, of spatial spots which we would like to calculate significant cell type cooccurences in}

\item{prob.th}{Probability threshold to convert the deconvolution probability matrix to binary presence/absence matrix}
}
\value{
a list containing the results table of cell type coocuurences
}
\description{
This function calculates the significant cooccuring cell types based on deconvoluted celltype probability matrix
}
